# fix errors in self-tests and the code they test. all now run through. (archive_builder and compression tests are disabled ATM)
also fixed some compile warnings inside cxxtest.

refs #117

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@4021 3db68df2-c116-0410-a063-a993310a9797

+++ libraries/cxxtest/include/cxxtest/TestListener.h
@@ -37,7 +37,7 @@ namespace CxxTest
         virtual void failedAssertSameData( const char * /*file*/, unsigned /*line*/,
                                            const char * /*xStr*/, const char * /*yStr*/,
                                            const char * /*sizeStr*/, const void * /*x*/,
-                                           const void * /*y*/, unsigned /*size*/ ) {}
+                                           const void * /*y*/, size_t /*size*/ ) {}
         virtual void failedAssertDelta( const char * /*file*/, unsigned /*line*/,
                                         const char * /*xStr*/, const char * /*yStr*/,
                                         const char * /*dStr*/, const char * /*x*/,
